Add td_api::writeGeneratedFilePart.
GitOrigin-RevId: 2dfcba1d5f14dfd5e701ac02d3a8fc96090a00f2

+++ tdutils/td/utils/filesystem.cpp
@@ -75,7 +75,7 @@ Status write_file(CSlice to, Slice data) {
   auto size = data.size();
   TRY_RESULT(to_file, FileFd::open(to, FileFd::Truncate | FileFd::Create | FileFd::Write));
   TRY_RESULT(written, to_file.write(data));
-  if (written != static_cast<size_t>(size)) {
+  if (written != size) {
     return Status::Error(PSLICE() << "Failed to write file: written " << written << " bytes instead of " << size);
   }
   to_file.close();
